sentiment_analysis_service/requirements.txt
blinker==1.9.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
dnspython==2.7.0
filelock==3.19.1
flask==3.1.2
fsspec==2025.9.0
hf-xet==1.1.9
huggingface-hub==0.34.4
idna==3.10
itsdangerous==2.2.0
jinja2==3.1.6
markupsafe==3.0.2
mpmath==1.3.0
networkx==3.5
numpy==2.3.2
packaging==25.0
pika==1.3.2
pymongo==4.14.1
python-dotenv==1.1.1
pyyaml==6.0.2
regex==2025.9.1
requests==2.32.5
safetensors==0.6.2
setuptools==80.9.0
sympy==1.14.0
tokenizers==0.22.0
torch==2.8.0
tqdm==4.67.1
transformers==4.56.0
typing-extensions==4.15.0
urllib3==2.5.0
werkzeug==3.1.3
